Split `MediaStore` trait from `EventCacheStore`. Rename `EventCacheStoreMedia` to `MediaStoreInner`. Move relevant tests into `MediaStoreIntegrationTests`.
