+++ crates/matrix-sdk-base/src/event_cache/store/media/integration_tests.rs
@@ -12,26 +12,29 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//! Trait and macro of integration tests for `EventCacheStoreMedia`
+//! Trait and macro of integration tests for `MediaStoreInner`
 //! implementations.
 
 use ruma::{
     events::room::MediaSource,
+    media::Method,
     mxc_uri, owned_mxc_uri,
     time::{Duration, SystemTime},
+    uint,
 };
 
-use super::{
-    EventCacheStoreMedia, MediaRetentionPolicy, media_service::IgnoreMediaRetentionPolicy,
+use super::{MediaRetentionPolicy, MediaStoreInner, media_service::IgnoreMediaRetentionPolicy};
+use crate::{
+    event_cache::store::media::media_service::MediaStore,
+    media::{MediaFormat, MediaRequestParameters, MediaThumbnailSettings},
 };
-use crate::media::{MediaFormat, MediaRequestParameters};
 
-/// [`EventCacheStoreMedia`] integration tests.
+/// [`MediaStoreInner`] integration tests.
 ///
 /// This trait is not meant to be used directly, but will be used with the
-/// `event_cache_store_media_integration_tests!` macro.
+/// `media_store_inner_integration_tests!` macro.
 #[allow(async_fn_in_trait)]
-pub trait EventCacheStoreMediaIntegrationTests {
+pub trait MediaStoreInnerIntegrationTests {
     /// Test media retention policy storage.
     async fn test_store_media_retention_policy(&self);
 
@@ -56,9 +59,9 @@ pub trait EventCacheStoreMediaIntegrationTests {
     async fn test_store_last_media_cleanup_time(&self);
 }
 
-impl<Store> EventCacheStoreMediaIntegrationTests for Store
+impl<Store> MediaStoreInnerIntegrationTests for Store
 where
-    Store: EventCacheStoreMedia + std::fmt::Debug,
+    Store: MediaStoreInner + std::fmt::Debug,
 {
     async fn test_store_media_retention_policy(&self) {
         let stored = self.media_retention_policy_inner().await.unwrap();
@@ -961,7 +964,7 @@ where
     }
 }
 
-/// Macro building to allow your [`EventCacheStoreMedia`] implementation to run
+/// Macro building to allow your [`MediaStoreInner`] implementation to run
 /// the entire tests suite locally.
 ///
 /// Can be run with the `with_media_size_tests` argument to include more tests
@@ -971,7 +974,7 @@ where
 ///
 /// You need to provide an `async fn get_event_cache_store() ->
 /// event_cache::store::Result<Store>` that provides a fresh event cache store
-/// that implements `EventCacheStoreMedia` on the same level you invoke the
+/// that implements `MediaStoreInner` on the same level you invoke the
 /// macro.
 ///
 /// ## Usage Example:
@@ -990,70 +993,323 @@ where
 ///         Ok(MyStore::new())
 ///     }
 ///
-///     event_cache_store_media_integration_tests!();
+///     media_store_inner_integration_tests!();
 /// }
 /// ```
 #[allow(unused_macros, unused_extern_crates)]
 #[macro_export]
-macro_rules! event_cache_store_media_integration_tests {
+macro_rules! media_store_inner_integration_tests {
     (with_media_size_tests) => {
-        mod event_cache_store_media_integration_tests {
-            $crate::event_cache_store_media_integration_tests!(@inner);
+        mod media_store_inner_integration_tests {
+            $crate::media_store_inner_integration_tests!(@inner);
 
             #[async_test]
             async fn test_media_max_file_size() {
-                let event_cache_store_media = get_event_cache_store().await.unwrap();
-                event_cache_store_media.test_media_max_file_size().await;
+                let media_store_inner = get_event_cache_store().await.unwrap();
+                media_store_inner.test_media_max_file_size().await;
             }
 
             #[async_test]
             async fn test_media_max_cache_size() {
-                let event_cache_store_media = get_event_cache_store().await.unwrap();
-                event_cache_store_media.test_media_max_cache_size().await;
+                let media_store_inner = get_event_cache_store().await.unwrap();
+                media_store_inner.test_media_max_cache_size().await;
             }
 
             #[async_test]
             async fn test_media_ignore_max_size() {
-                let event_cache_store_media = get_event_cache_store().await.unwrap();
-                event_cache_store_media.test_media_ignore_max_size().await;
+                let media_store_inner = get_event_cache_store().await.unwrap();
+                media_store_inner.test_media_ignore_max_size().await;
             }
         }
     };
 
     () => {
-        mod event_cache_store_media_integration_tests {
-            $crate::event_cache_store_media_integration_tests!(@inner);
+        mod media_store_inner_integration_tests {
+            $crate::media_store_inner_integration_tests!(@inner);
         }
     };
 
     (@inner) => {
         use matrix_sdk_test::async_test;
-        use $crate::event_cache::store::media::EventCacheStoreMediaIntegrationTests;
+        use $crate::event_cache::store::media::MediaStoreInnerIntegrationTests;
 
         use super::get_event_cache_store;
 
         #[async_test]
         async fn test_store_media_retention_policy() {
-            let event_cache_store_media = get_event_cache_store().await.unwrap();
-            event_cache_store_media.test_store_media_retention_policy().await;
+            let media_store_inner = get_event_cache_store().await.unwrap();
+            media_store_inner.test_store_media_retention_policy().await;
         }
 
         #[async_test]
         async fn test_media_expiry() {
-            let event_cache_store_media = get_event_cache_store().await.unwrap();
-            event_cache_store_media.test_media_expiry().await;
+            let media_store_inner = get_event_cache_store().await.unwrap();
+            media_store_inner.test_media_expiry().await;
         }
 
         #[async_test]
         async fn test_media_ignore_expiry() {
-            let event_cache_store_media = get_event_cache_store().await.unwrap();
-            event_cache_store_media.test_media_ignore_expiry().await;
+            let media_store_inner = get_event_cache_store().await.unwrap();
+            media_store_inner.test_media_ignore_expiry().await;
         }
 
         #[async_test]
         async fn test_store_last_media_cleanup_time() {
-            let event_cache_store_media = get_event_cache_store().await.unwrap();
-            event_cache_store_media.test_store_last_media_cleanup_time().await;
+            let media_store_inner = get_event_cache_store().await.unwrap();
+            media_store_inner.test_store_last_media_cleanup_time().await;
+        }
+    };
+}
+
+/// [`MediaStore`] integration tests.
+///
+/// This trait is not meant to be used directly, but will be used with the
+/// `media_store_inner_integration_tests!` macro.
+#[allow(async_fn_in_trait)]
+pub trait MediaStoreIntegrationTests {
+    /// Test media content storage.
+    async fn test_media_content(&self);
+
+    /// Test replacing a MXID.
+    async fn test_replace_media_key(&self);
+}
+
+impl<Store> MediaStoreIntegrationTests for Store
+where
+    Store: MediaStore + std::fmt::Debug,
+{
+    async fn test_media_content(&self) {
+        let uri = mxc_uri!("mxc://localhost/media");
+        let request_file = MediaRequestParameters {
+            source: MediaSource::Plain(uri.to_owned()),
+            format: MediaFormat::File,
+        };
+        let request_thumbnail = MediaRequestParameters {
+            source: MediaSource::Plain(uri.to_owned()),
+            format: MediaFormat::Thumbnail(MediaThumbnailSettings::with_method(
+                Method::Crop,
+                uint!(100),
+                uint!(100),
+            )),
+        };
+
+        let other_uri = mxc_uri!("mxc://localhost/media-other");
+        let request_other_file = MediaRequestParameters {
+            source: MediaSource::Plain(other_uri.to_owned()),
+            format: MediaFormat::File,
+        };
+
+        let content: Vec<u8> = "hello".into();
+        let thumbnail_content: Vec<u8> = "world".into();
+        let other_content: Vec<u8> = "foo".into();
+
+        // Media isn't present in the cache.
+        assert!(
+            self.get_media_content(&request_file).await.unwrap().is_none(),
+            "unexpected media found"
+        );
+        assert!(
+            self.get_media_content(&request_thumbnail).await.unwrap().is_none(),
+            "media not found"
+        );
+
+        // Let's add the media.
+        self.add_media_content(&request_file, content.clone(), IgnoreMediaRetentionPolicy::No)
+            .await
+            .expect("adding media failed");
+
+        // Media is present in the cache.
+        assert_eq!(
+            self.get_media_content(&request_file).await.unwrap().as_ref(),
+            Some(&content),
+            "media not found though added"
+        );
+        assert_eq!(
+            self.get_media_content_for_uri(uri).await.unwrap().as_ref(),
+            Some(&content),
+            "media not found by URI though added"
+        );
+
+        // Let's remove the media.
+        self.remove_media_content(&request_file).await.expect("removing media failed");
+
+        // Media isn't present in the cache.
+        assert!(
+            self.get_media_content(&request_file).await.unwrap().is_none(),
+            "media still there after removing"
+        );
+        assert!(
+            self.get_media_content_for_uri(uri).await.unwrap().is_none(),
+            "media still found by URI after removing"
+        );
+
+        // Let's add the media again.
+        self.add_media_content(&request_file, content.clone(), IgnoreMediaRetentionPolicy::No)
+            .await
+            .expect("adding media again failed");
+
+        assert_eq!(
+            self.get_media_content(&request_file).await.unwrap().as_ref(),
+            Some(&content),
+            "media not found after adding again"
+        );
+
+        // Let's add the thumbnail media.
+        self.add_media_content(
+            &request_thumbnail,
+            thumbnail_content.clone(),
+            IgnoreMediaRetentionPolicy::No,
+        )
+        .await
+        .expect("adding thumbnail failed");
+
+        // Media's thumbnail is present.
+        assert_eq!(
+            self.get_media_content(&request_thumbnail).await.unwrap().as_ref(),
+            Some(&thumbnail_content),
+            "thumbnail not found"
+        );
+
+        // We get a file with the URI, we don't know which one.
+        assert!(
+            self.get_media_content_for_uri(uri).await.unwrap().is_some(),
+            "media not found by URI though two where added"
+        );
+
+        // Let's add another media with a different URI.
+        self.add_media_content(
+            &request_other_file,
+            other_content.clone(),
+            IgnoreMediaRetentionPolicy::No,
+        )
+        .await
+        .expect("adding other media failed");
+
+        // Other file is present.
+        assert_eq!(
+            self.get_media_content(&request_other_file).await.unwrap().as_ref(),
+            Some(&other_content),
+            "other file not found"
+        );
+        assert_eq!(
+            self.get_media_content_for_uri(other_uri).await.unwrap().as_ref(),
+            Some(&other_content),
+            "other file not found by URI"
+        );
+
+        // Let's remove media based on URI.
+        self.remove_media_content_for_uri(uri).await.expect("removing all media for uri failed");
+
+        assert!(
+            self.get_media_content(&request_file).await.unwrap().is_none(),
+            "media wasn't removed"
+        );
+        assert!(
+            self.get_media_content(&request_thumbnail).await.unwrap().is_none(),
+            "thumbnail wasn't removed"
+        );
+        assert!(
+            self.get_media_content(&request_other_file).await.unwrap().is_some(),
+            "other media was removed"
+        );
+        assert!(
+            self.get_media_content_for_uri(uri).await.unwrap().is_none(),
+            "media found by URI wasn't removed"
+        );
+        assert!(
+            self.get_media_content_for_uri(other_uri).await.unwrap().is_some(),
+            "other media found by URI was removed"
+        );
+    }
+
+    async fn test_replace_media_key(&self) {
+        let uri = mxc_uri!("mxc://sendqueue.local/tr4n-s4ct-10n1-d");
+        let req = MediaRequestParameters {
+            source: MediaSource::Plain(uri.to_owned()),
+            format: MediaFormat::File,
+        };
+
+        let content = "hello".as_bytes().to_owned();
+
+        // Media isn't present in the cache.
+        assert!(self.get_media_content(&req).await.unwrap().is_none(), "unexpected media found");
+
+        // Add the media.
+        self.add_media_content(&req, content.clone(), IgnoreMediaRetentionPolicy::No)
+            .await
+            .expect("adding media failed");
+
+        // Sanity-check: media is found after adding it.
+        assert_eq!(self.get_media_content(&req).await.unwrap().unwrap(), b"hello");
+
+        // Replacing a media request works.
+        let new_uri = mxc_uri!("mxc://matrix.org/tr4n-s4ct-10n1-d");
+        let new_req = MediaRequestParameters {
+            source: MediaSource::Plain(new_uri.to_owned()),
+            format: MediaFormat::File,
+        };
+        self.replace_media_key(&req, &new_req)
+            .await
+            .expect("replacing the media request key failed");
+
+        // Finding with the previous request doesn't work anymore.
+        assert!(
+            self.get_media_content(&req).await.unwrap().is_none(),
+            "unexpected media found with the old key"
+        );
+
+        // Finding with the new request does work.
+        assert_eq!(self.get_media_content(&new_req).await.unwrap().unwrap(), b"hello");
+    }
+}
+
+/// Macro building to allow your [`MediaStore`] implementation to run
+/// the entire tests suite locally.
+///
+/// You need to provide an `async fn get_media_store() ->
+/// event_cache::store::media::Result<Store>` that provides a fresh media store
+/// that implements `MediaStoreInner` on the same level you invoke the
+/// macro.
+///
+/// ## Usage Example:
+/// ```no_run
+/// # use matrix_sdk_base::event_cache::store::media{
+/// #    MediaStore,
+/// #    MemoryStore as MyStore,
+/// #    Result as MediaStoreResult,
+/// # };
+///
+/// #[cfg(test)]
+/// mod tests {
+///     use super::{MediaStoreResult, MyStore};
+///
+///     async fn get_media_store() -> MediaStoreResult<MyStore> {
+///         Ok(MyStore::new())
+///     }
+///
+///     media_store_integration_tests!();
+/// }
+/// ```
+#[allow(unused_macros, unused_extern_crates)]
+#[macro_export]
+macro_rules! media_store_integration_tests {
+    () => {
+        mod media_store_integration_tests {
+            use matrix_sdk_test::async_test;
+
+            use super::get_media_store;
+
+            #[async_test]
+            async fn test_media_content() {
+                let event_cache_store = get_media_store().await.unwrap().into_event_cache_store();
+                event_cache_store.test_media_content().await;
+            }
+
+            #[async_test]
+            async fn test_replace_media_key() {
+                let event_cache_store = get_media_store().await.unwrap().into_event_cache_store();
+                event_cache_store.test_replace_media_key().await;
+            }
         }
     };
 }
