+++ crates/matrix-sdk-base/src/event_cache/store/traits.rs
@@ -22,16 +22,10 @@ use matrix_sdk_common::{
         RawChunk, Update,
     },
 };
-use ruma::{EventId, MxcUri, OwnedEventId, RoomId, events::relation::RelationType};
+use ruma::{EventId, OwnedEventId, RoomId, events::relation::RelationType};
 
-use super::{
-    EventCacheStoreError,
-    media::{IgnoreMediaRetentionPolicy, MediaRetentionPolicy},
-};
-use crate::{
-    event_cache::{Event, Gap},
-    media::MediaRequestParameters,
-};
+use super::EventCacheStoreError;
+use crate::event_cache::{Event, Gap};
 
 /// A default capacity for linked chunks, when manipulating in conjunction with
 /// an `EventCacheStore` implementation.
@@ -169,129 +163,6 @@ pub trait EventCacheStore: AsyncTraitDeps {
     /// If the event was already stored with the same id, it must be replaced,
     /// without causing an error.
     async fn save_event(&self, room_id: &RoomId, event: Event) -> Result<(), Self::Error>;
-
-    /// Add a media file's content in the media store.
-    ///
-    /// # Arguments
-    ///
-    /// * `request` - The `MediaRequest` of the file.
-    ///
-    /// * `content` - The content of the file.
-    async fn add_media_content(
-        &self,
-        request: &MediaRequestParameters,
-        content: Vec<u8>,
-        ignore_policy: IgnoreMediaRetentionPolicy,
-    ) -> Result<(), Self::Error>;
-
-    /// Replaces the given media's content key with another one.
-    ///
-    /// This should be used whenever a temporary (local) MXID has been used, and
-    /// it must now be replaced with its actual remote counterpart (after
-    /// uploading some content, or creating an empty MXC URI).
-    ///
-    /// ⚠ No check is performed to ensure that the media formats are consistent,
-    /// i.e. it's possible to update with a thumbnail key a media that was
-    /// keyed as a file before. The caller is responsible of ensuring that
-    /// the replacement makes sense, according to their use case.
-    ///
-    /// This should not raise an error when the `from` parameter points to an
-    /// unknown media, and it should silently continue in this case.
-    ///
-    /// # Arguments
-    ///
-    /// * `from` - The previous `MediaRequest` of the file.
-    ///
-    /// * `to` - The new `MediaRequest` of the file.
-    async fn replace_media_key(
-        &self,
-        from: &MediaRequestParameters,
-        to: &MediaRequestParameters,
-    ) -> Result<(), Self::Error>;
-
-    /// Get a media file's content out of the media store.
-    ///
-    /// # Arguments
-    ///
-    /// * `request` - The `MediaRequest` of the file.
-    async fn get_media_content(
-        &self,
-        request: &MediaRequestParameters,
-    ) -> Result<Option<Vec<u8>>, Self::Error>;
-
-    /// Remove a media file's content from the media store.
-    ///
-    /// # Arguments
-    ///
-    /// * `request` - The `MediaRequest` of the file.
-    async fn remove_media_content(
-        &self,
-        request: &MediaRequestParameters,
-    ) -> Result<(), Self::Error>;
-
-    /// Get a media file's content associated to an `MxcUri` from the
-    /// media store.
-    ///
-    /// In theory, there could be several files stored using the same URI and a
-    /// different `MediaFormat`. This API is meant to be used with a media file
-    /// that has only been stored with a single format.
-    ///
-    /// If there are several media files for a given URI in different formats,
-    /// this API will only return one of them. Which one is left as an
-    /// implementation detail.
-    ///
-    /// # Arguments
-    ///
-    /// * `uri` - The `MxcUri` of the media file.
-    async fn get_media_content_for_uri(&self, uri: &MxcUri)
-    -> Result<Option<Vec<u8>>, Self::Error>;
-
-    /// Remove all the media files' content associated to an `MxcUri` from the
-    /// media store.
-    ///
-    /// This should not raise an error when the `uri` parameter points to an
-    /// unknown media, and it should return an Ok result in this case.
-    ///
-    /// # Arguments
-    ///
-    /// * `uri` - The `MxcUri` of the media files.
-    async fn remove_media_content_for_uri(&self, uri: &MxcUri) -> Result<(), Self::Error>;
-
-    /// Set the `MediaRetentionPolicy` to use for deciding whether to store or
-    /// keep media content.
-    ///
-    /// # Arguments
-    ///
-    /// * `policy` - The `MediaRetentionPolicy` to use.
-    async fn set_media_retention_policy(
-        &self,
-        policy: MediaRetentionPolicy,
-    ) -> Result<(), Self::Error>;
-
-    /// Get the current `MediaRetentionPolicy`.
-    fn media_retention_policy(&self) -> MediaRetentionPolicy;
-
-    /// Set whether the current [`MediaRetentionPolicy`] should be ignored for
-    /// the media.
-    ///
-    /// The change will be taken into account in the next cleanup.
-    ///
-    /// # Arguments
-    ///
-    /// * `request` - The `MediaRequestParameters` of the file.
-    ///
-    /// * `ignore_policy` - Whether the current `MediaRetentionPolicy` should be
-    ///   ignored.
-    async fn set_ignore_media_retention_policy(
-        &self,
-        request: &MediaRequestParameters,
-        ignore_policy: IgnoreMediaRetentionPolicy,
-    ) -> Result<(), Self::Error>;
-
-    /// Clean up the media cache with the current `MediaRetentionPolicy`.
-    ///
-    /// If there is already an ongoing cleanup, this is a noop.
-    async fn clean_up_media_cache(&self) -> Result<(), Self::Error>;
 }
 
 #[repr(transparent)]
@@ -389,71 +260,6 @@ impl<T: EventCacheStore> EventCacheStore for EraseEventCacheStoreError<T> {
 
     async fn save_event(&self, room_id: &RoomId, event: Event) -> Result<(), Self::Error> {
         self.0.save_event(room_id, event).await.map_err(Into::into)
-    }
-
-    async fn add_media_content(
-        &self,
-        request: &MediaRequestParameters,
-        content: Vec<u8>,
-        ignore_policy: IgnoreMediaRetentionPolicy,
-    ) -> Result<(), Self::Error> {
-        self.0.add_media_content(request, content, ignore_policy).await.map_err(Into::into)
-    }
-
-    async fn replace_media_key(
-        &self,
-        from: &MediaRequestParameters,
-        to: &MediaRequestParameters,
-    ) -> Result<(), Self::Error> {
-        self.0.replace_media_key(from, to).await.map_err(Into::into)
-    }
-
-    async fn get_media_content(
-        &self,
-        request: &MediaRequestParameters,
-    ) -> Result<Option<Vec<u8>>, Self::Error> {
-        self.0.get_media_content(request).await.map_err(Into::into)
-    }
-
-    async fn remove_media_content(
-        &self,
-        request: &MediaRequestParameters,
-    ) -> Result<(), Self::Error> {
-        self.0.remove_media_content(request).await.map_err(Into::into)
-    }
-
-    async fn get_media_content_for_uri(
-        &self,
-        uri: &MxcUri,
-    ) -> Result<Option<Vec<u8>>, Self::Error> {
-        self.0.get_media_content_for_uri(uri).await.map_err(Into::into)
-    }
-
-    async fn remove_media_content_for_uri(&self, uri: &MxcUri) -> Result<(), Self::Error> {
-        self.0.remove_media_content_for_uri(uri).await.map_err(Into::into)
-    }
-
-    async fn set_media_retention_policy(
-        &self,
-        policy: MediaRetentionPolicy,
-    ) -> Result<(), Self::Error> {
-        self.0.set_media_retention_policy(policy).await.map_err(Into::into)
-    }
-
-    fn media_retention_policy(&self) -> MediaRetentionPolicy {
-        self.0.media_retention_policy()
-    }
-
-    async fn set_ignore_media_retention_policy(
-        &self,
-        request: &MediaRequestParameters,
-        ignore_policy: IgnoreMediaRetentionPolicy,
-    ) -> Result<(), Self::Error> {
-        self.0.set_ignore_media_retention_policy(request, ignore_policy).await.map_err(Into::into)
-    }
-
-    async fn clean_up_media_cache(&self) -> Result<(), Self::Error> {
-        self.0.clean_up_media_cache().await.map_err(Into::into)
     }
 }
